+++ inc/em_configuration.h
@@ -189,17 +189,17 @@ public:
             return em_crypto_t::platform_compute_shared_secret(secret, secret_len, remote_pub, pub_len, local_priv, static_cast<uint8_t>(priv_len));
     }
 
-    int compute_digest(unsigned char num, unsigned char **addr, unsigned int *len, unsigned char *digest) {
+    int compute_digest(unsigned char num, unsigned char **addr, size_t *len, unsigned char *digest) {
         return em_crypto_t::platform_SHA256(num, addr, len, digest); 
     }
 
     int compute_kdk(unsigned char *key, unsigned short keylen, 
         unsigned char num_elem, unsigned char **addr, 
-        unsigned int *len, unsigned char *hmac) {
+        size_t *len, unsigned char *hmac) {
             return em_crypto_t::platform_hmac_SHA256(key, keylen, num_elem, addr, len, hmac);
     }
 
-    int derive_key(unsigned char *key, unsigned char *label_prefix, unsigned int label_prefix_len, 
+    int derive_key(unsigned char *key, unsigned char *label_prefix, size_t label_prefix_len, 
         char *label, unsigned char *res, unsigned int res_len) {
             return em_crypto_t::wps_key_derivation_function(key, label_prefix, label_prefix_len, label, res, res_len);
     }
